+++ c3smembership/business/payment_information.py
@@ -22,14 +22,15 @@ class PaymentInformation(object):
 
     # pylint: disable=too-many-arguments
     def get_payments(
-            self, page_number, page_size, sort_property='date',
+            self, page_number=None, page_size=None, sort_property='date',
             sort_direction='asc', from_date=None, to_date=None):
         """
         Gets the payments for a page filtered by dates.
 
         Args:
-            page_number: The number of the page of payments to be returned.
-            page_size: The size of the pages of payments.
+            page_number: Optional. The number of the page of payments to be
+                returned.
+            page_size: Optional. The size of the pages of payments.
             sort_property: Optional. A string representing the payment property
                 by which the payment list is sorted. Valid sort properties are:
 
@@ -57,3 +58,20 @@ class PaymentInformation(object):
             sort_direction,
             from_date,
             to_date)
+
+    def get_payment_count(self, from_date=None, to_date=None):
+        """
+        Gets the count of payments of which the payment date is not older than
+        from date and not younger than to date.
+
+        Args:
+            from_date: Optional. A datetime.date specifying the oldest payment
+                date for payments to be counted.
+            to_date: Optional. A datetie.date specifying the youngest payment
+                date for payments to be counted.
+
+        Return:
+            An integer representing the count of payments available not older
+            than from date and not younger than to date.
+        """
+        return self._payment_repository.get_payment_count(from_date, to_date)